Rag/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
dataclasses-json==0.6.7
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langsmith==0.4.4
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.1
orjson==3.10.18
packaging==24.2
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pypdf==5.7.0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
zstandard==0.23.0
